- fixed inconsistent data type used for protocol (int and unsigned short);   All fields carrying protocol or port moved unsigned short. More safer and   many issues related to casting eliminated.
git-svn-id: https://openser.svn.sourceforge.net/svnroot/openser/trunk@1579 689a6050-402a-0410-94f2-e92a70836424

+++ modules_k/nathelper/nathelper.c
@@ -2098,7 +2098,7 @@ nh_timer(unsigned int ticks, void *param)
 	struct hostent* he;
 	struct socket_info* send_sock;
 	unsigned int flags;
-	int proto;
+	unsigned short proto;
 	str opt;
 
 	buf = NULL;
